requirements.txt
GOOGLE_API_KEY
streamlit
google-generativeai
python-dotenv